requirements.txt
pandas
pyyaml
opencv-python
matplotlib
shapely
scikit-image
scipy
sam2
ultralytics